Build 32b code on Darwin WAR problems with __align__ on gcc 4.2 and below
